moved partition session storage
